Feature/json algorithm output (#1)
I created a SUPER basic server to test the endpoint. I also changed the way the json output is getting build.

Broke out the server and inetaccess into 2 different files

* Updated to clean up code, took out print statements, returning cleaner json

Updated so the http response is returned regardless if it's a 200 or not BUT it's only counted towards the scoring if the response is a 200.